requirements.txt
flask
flask_session
google-generativeai
python-dotenv
boto3
pdfkit
